cleanup home page
